+++ src/main/java/ru/otus/library/shared/exceptions/author/AuthorNotFoundByIdException.java
@@ -3,5 +3,5 @@ package ru.otus.library.shared.exceptions.author;
 import ru.otus.library.shared.exceptions.EntityNotFoundException;
 
 public class AuthorNotFoundByIdException extends EntityNotFoundException {
-    public AuthorNotFoundByIdException(Integer id) { super(id.toString()); }
+    public AuthorNotFoundByIdException(Long id) { super(id.toString()); }
 }
