+++ src/main/java/ru/otus/library/shared/exceptions/book/BookNotFoundByIdException.java
@@ -3,5 +3,5 @@ package ru.otus.library.shared.exceptions.book;
 import ru.otus.library.shared.exceptions.EntityNotFoundException;
 
 public class BookNotFoundByIdException extends EntityNotFoundException {
-    public BookNotFoundByIdException(Integer id) { super(id.toString()); }
+    public BookNotFoundByIdException(Long id) { super(id.toString()); }
 }
