+++ src/main/java/ru/otus/library/shared/exceptions/genre/GenreNotFoundByIdException.java
@@ -3,5 +3,5 @@ package ru.otus.library.shared.exceptions.genre;
 import ru.otus.library.shared.exceptions.EntityNotFoundException;
 
 public class GenreNotFoundByIdException extends EntityNotFoundException {
-    public GenreNotFoundByIdException(Integer id) { super(id.toString()); }
+    public GenreNotFoundByIdException(Long id) { super(id.toString()); }
 }
